rename: useGetMyScheduleQuery 훅을 schedule 관련 파일로 이동 (meeting → schedule)
