+++ frontend/src/stores/servers/schedule/queries.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { handleGetMeetingSchedules } from '@apis/schedules';
+import { getMeetingMySchedule, handleGetMeetingSchedules } from '@apis/schedules';
 
 import { QUERY_KEY } from '@constants/queryKeys';
 
@@ -8,5 +8,12 @@ export const useGetSchedules = (uuid: string, selectedAttendee: string) =>
   useQuery({
     queryKey: [QUERY_KEY.meetingSchedules, selectedAttendee],
     queryFn: () => handleGetMeetingSchedules({ uuid, attendeeName: selectedAttendee }),
+    staleTime: 0,
+  });
+
+export const useGetMyScheduleQuery = (uuid: string, userName: string) =>
+  useQuery({
+    queryKey: [QUERY_KEY.meetingMySchedule, { userName }],
+    queryFn: () => getMeetingMySchedule(uuid),
     staleTime: 0,
   });
